ExyteBot/requirements.txt
aiogram==2.11.2
psycopg2-binary==2.8.6
requests==2.25.1
seaborn==0.11.2
pandas==1.1.5